Update Section 2, Lesson 3 to remind users to redeploy contract after change
I noticed that some folks in the discord were having issues as Section 2, Lesson 2 asks them to deploy to Rinkeby while minting a character and running attacks in the deploy.js file.

In the next lesson (Section 2, Lesson 3), we add the utility functions to MyEpicGame.sol but we do not explicitly call out a redeploy of the contract. If the user does not do this, they run into errors because 
- the utility functions do not exist (they did not redeploy the contract)
- they already have an NFT (due to minting in the previous deploy)

This PR just adds a bit at the end that reminds them to deploy their contract again and includes a sample deploy.js file to aid them.