Solana validator is more accurate than miner.
I understand wanting to relate things to the ETH world, but in this case I think there isn't a direct translation and could be misleading. In PoS, validators don't mine. Even in ETH 2.0 they'll move over to a "stake" or "validate" terminology so this will probably get more confusing over time.

https://ethereum.org/en/developers/docs/consensus-mechanisms/pos/