Translation of Lesson 1
Hello! Saw that you guys were planning to make the projects available in Spanish, so I hope I can help a little bit.

The parts that wanted people to interact in the Discord could be deleted, because the servers is mainly in English. Unless there are some plans to open an Spanish Discord for Buildspace or maybe some text channels in Spanish, who knows.

Hope it helps!