Add missing `const presale` when updating `getCandyMachineState`
In this lesson we write `getCandyMachineState` twice - the first time just to fetch + log the data, the second time also to call `setCandyMachine`. In the second time the `const presale = ...` section was missing. We reference `presale` in `setCandyMachine`, so we need it in this function. This commit just copies the `const presale = ...` from the first version of the function to the second, fixing the error.