removed partially implemented logic
I'm guessing this logic may have been introduced recently.

it is missing a couple of supporting parts for the arena index.js. 
i.e. currentAccount is not declared in this scope and the logic to set it on page render is also missing. 

If I've not misunderstood, I'm guessing the preferred solution is to add the missing logic supporting the lines removed.