Adding ABI variable
I went through this section and had trouble getting things to work because I had copy/pasted this code and it overwrote the ABI variable I had in place previously. It took me a while to figure this out and I'm assuming others might have the same issue. Adding this in for clarity.