Update Lesson_4_Sell_Domains_As_NFTs.md
Whenever i see a comment (//) on buildspace, i immediately know that a particular code is an addition. When i reached section  "sell_domain_as_nfts", i was confused if it was a new "register" function or was the old function modified? that is why i thought to add "payable" as a comment so that new programmers won't have this confusion.